### 2025-04-06 #8010 @editorialbot
[PRE REVIEW]: py_neuromodulation: Signal processing and decoding for neural electrophysiological recordings
**Submitting author:** @timonmerk (<a href="http://orcid.org/0000-0003-3011-2612">Timon Merk</a>)
**Repository:** https://github.com/neuromodulation/py_neuromodulation
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.1.2
