### 2025-04-13 7f24775
Another attempt by doing subscription with plotly.react() instead of plotly.restyle()
