### 2025-05-15 #8010 @timonmerk
@sappelhoff and @britta-wstnr I am very sorry to tag you here directly. But I just wondered about the usual turn-around time of the editors? It would be very helpful to know if the paper was considered for this journal? Many thanks!
