### 2025-05-15 #8010 @sappelhoff
Hey @timonmerk, no worries about the tag. We currently have a very big backlog in papers that need to be edited, so things are going a bit slower than they used to. Sorry for that!

As for this specific paper, @Kevin-Mattheus-Moerman is the right _aeic_ to reach out to. He will find an appropriate editor for this submission.

As for taking this paper over as an editor myself, I am not sure whether I am the right person (conflict of interest: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)

@Kevin-Mattheus-Moerman: @timonmerk and I have (to the best of my memory) met personally on one occasion and have collaborated on software such as mne-bids (https://github.com/mne-tools/mne-bids/pull/1051). @timonmerk has also once invited me as a speaker for an event that he organized. Other than that I do not think we have associations that fall under our COI policy. I would like to leave it up to you to waive my COI, or to find another editor.
