### 2025-05-19 #8010 @jromanowska
Hi, @timonmerk  :wave:

I will take care of the process from now on. I will try to find reviewers - thanks for your suggestions above.

As the first task, I'd like you to make sure that your paper has a "Statement of need" section (see [JOSS docs](https://joss.readthedocs.io/en/latest/paper.html) and [editorialbot's comment](https://github.com/openjournals/joss-reviews/issues/8010#issuecomment-2781275785) ) and maybe you can consider cutting down a bit on the length.
