+++ paper/paper.md
@@ -60,20 +60,18 @@ Invasive brain signal decoding can revolutionize the clinical utility of neurote
 
 Recently, deep learning foundation models were presented that showed a performance leap on many clinical relevant downstream tasks [@yuan_brainwave_2024;@li_deep_2025]. Neural signal processing for machine learning and medical device engineering is a critical step for the advancement of the fields of neural decoding and brain computer interfaces. We, therefore, developed a neural signal processing, feature estimation and decoding toolbox [py_neuromodulation](https://github.com/neuromodulation/py_neuromodulation), that includes several established and standardized pipelines that simplify processing of neural recordings.  
 Analysis of electrophysiological recordings is commonly started by multiple preprocessing steps: resampling, notch-filtering, normalization, re-referencing (subtraction of neighboring channels), filtering (lowpass, highpass, bandpass), or artifact rejection [@cohen_analyzing_2014]. Next, the signals or features of interest are computed. Most commonly, signals are analyzed in the spectral domain. Here, different methods can be utilized, including Fast Fourier Transform, Welch transform, bandpass filtering, and others. In addition to spectral features, different feature modalities such as waveform-shape [@cole_brain_2017], periodic and aperiodic spectral parametrization [@donoghue_parameterizing_2020], burst features [@tinkhauser_beta_2018], and others can be computed.
-Many toolboxes exist already that enable analysis of electrophysiological recordings: in Matlab Fieldtrip [@oostenveld_fieldtrip_2011] or Brainstorm [@tadel_brainstorm_2011], and in Python mne-python [@gramfort_meg_2013]. These include extensive methods for various processing and feature computation.
-By introducing *py_neuromodulation*, we want to address challenges in electrophysiological signal analysis, specifically in the context of invasive brain signal decoding.
-First, the variety of different processing parameters poses a challenge first to select optimal methods for machine learning-based decoding, and second, hinders reproducibility, since often necessary parameters are not reported in scientific literature or reproducible from the derived data. Here, we introduce parametrization files, *settings.yaml* and *channels.tsv*, that include all required steps for reproducibility. Additionally, it allows for a streamlined pipeline of pre-processing, feature estimation and decoding. Thus, the performance contribution of various pre-processing steps can be assessed in combination with different feature estimation modalities and machine learning model architectures.  
-Next, offline analysis often includes whole-recording normalization, non-causal filtering, or feature computation with test-set data leakage [@merk_machine_2022], which limits real-time compatibility. In py_neuromodulation, the data-source is represented as a Python generator, and can be interchangeable for online and offline processing, meaning that there is no difference in handling data streamed directly from a brain implant, or stored from an archival dataset. This enables training of generalizable models that are directly applicable to online applications, using exactly the same algorithms. 
-Finally, electrophysiological research entails multiple technical and conceptual hurdles that pose substantial entry barriers for scientists without software development training. By providing an intuitive GUI for parametrization and raw-data and feature visualization, we aim to lower the domain knowledge requirements for entering this field.
-Notably, py_neuromodulation was already used in several scientific projects with various applications: ECoG and DBS-LFP movement decoding [@merk_electrocorticography_2022;@cavallo_reinforcement_2025], movement intention classification [@kohler_dopamine_2024], seizure classification [@merk_invasive_2023], emotional valence decoding [@merk_invasive_2023], gait symptom estimation in rodents [@elgarulli_elgarullineurokin_2025], amongst others.
 
[…]
