### 2025-05-21 #8010 @jromanowska
:wave: @Julie-Fabre, @danielmk, @drcandacemakedamoore, @finsberg, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
