### 2025-05-23 #8258 @danielmk
:wave: @timonmerk I am one of the reviewers of your JOSS submission. I will soon start working through the above checklist. Once I have completed the checklist, I will recommend publication of your paper. If I encounter any issues with completing the checklist, I will open an issue on your github repository and we will figure out how to resolve the issue.

However, the review is also supposed to offer the authors constructive feedback to make their software as useful as possible. While working through the checklist, I might discover things that I feel could be improved, but that are not necessary for completing the checklist. I will explicitly mark issues that are related to the checklist and I will mark issues that are unrelated to the checklist. Feel free to ask me if you are unsure about an issue's priority. I will get to work on the repository/checklist next week. Have a nice weekend.
