### 2025-05-27 neuromodulation/py_neuromodulation#391 @danielmk
[JOSS Review] Checklist: Contribution and authorship clarification
This issue is part of the JOSS review process: https://github.com/openjournals/joss-reviews/issues/8258

It relates to the checklist point "Contribution and authorship".

Authorship is a complex issue and ultimately the authors decide which contributions are sufficient for authorship. I just want to make sure there are no errors in the current author list. I noticed some contributors, as shown on the github page are not on the author list. Some of these seem to have made substantial contributions, judging purely by the number of commits and lines of code. Please verify that these contributions are indeed insufficient to constitute authorship.

If you are confident that the authorship list as seen in [this version of the document](https://raw.githubusercontent.com/openjournals/joss-papers/joss.08258/joss.08258/10.21105.joss.08258.pdf) is correct, you can close this issue.
