### 2025-05-27 neuromodulation/py_neuromodulation#392 @danielmk
[JOSS Review] Checklist: Functionality - pytest failing
This issue is part of the JOSS review process: https://github.com/openjournals/joss-reviews/issues/8258

Running pytest in \py_neuromodulation\tests, I got the following error:
```
ImportError while loading conftest 'C:\Users\Daniel\repos\py_neuromodulation\tests\conftest.py'.
conftest.py:4: in <module>
    import py_neuromodulation as nm
..\py_neuromodulation\__init__.py:73: in <module>
    from .stream.stream import Stream
..\py_neuromodulation\stream\__init__.py:3: in <module>
    from .stream import Stream
..\py_neuromodulation\stream\stream.py:13: in <module>
    from py_neuromodulation.utils.file_writer import MsgPackFileWriter
..\py_neuromodulation\utils\file_writer.py:1: in <module>
    import msgpack
E   ModuleNotFoundError: No module named 'msgpack'
```

I installed with pip as per instructions. The code snippet under "Basic Usage" works. Python is 3.10.16 I'm on Windows 10. On the main branch.
