### 2025-05-27 5938c3d
add pyqt and msgpack missing dependencies
