+++ pyproject.toml
@@ -59,9 +59,11 @@ dependencies = [
   "fooof",
   "nolds >=0.6.1",
   "pybispectra>=1.2.0",
-  "numba >= 0.60.0", # Required to avoid uv version conflicts with pybispectra
+  "numba >= 0.60.0",    # Required to avoid uv version conflicts with pybispectra
   # Graphing
   "matplotlib>=3.9.0",
+  "PyQt6",             # necessary for plotting
+  "msgpack",
   "seaborn >= 0.11",
   "scikit-optimize",
   "mrmr_selection",
