### 2025-05-27 neuromodulation/py_neuromodulation#393 @finsberg
[JOSS REVIEW] Links in README does not work
Hi,

The links under basic usage goes to the page https://py-neuromodulation.readthedocs.io/en/latest/usage.htm https://py-neuromodulation.readthedocs.io/en/latest/auto_examples/plot_0_first_demo.html which doesn't exist.

This issue is related to the review at https://github.com/openjournals/joss-reviews/issues/8258
