### 2025-05-27 59ddb3c
Update README.rst
Fix broken links
