### 2025-05-27 neuromodulation/py_neuromodulation#393 @timonmerk
Thank you for that remark. It was fixed now and replaced by the correct links.
