+++ _sources/auto_examples/plot_2_example_add_feature.rst.txt
@@ -1,0 +1,447 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "auto_examples/plot_2_example_add_feature.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        :ref:`Go to the end <sphx_glr_download_auto_examples_plot_2_example_add_feature.py>`
+        to download the full example code.
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_auto_examples_plot_2_example_add_feature.py:
+
+
+===================
+Adding New Features
+===================
+
+.. GENERATED FROM PYTHON SOURCE LINES 9-13
+
+.. code-block:: Python
+
+    import py_neuromodulation as nm
+    import numpy as np
+    from typing import Iterable
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 14-17
+
+In this example we will demonstrate how a new feature can be added to the existing feature pipeline.
+This can be done by creating a new feature class that implements the protocol class :class:`~features.NMFeature`
+and registering it with the :func:`~features.AddCustomFeature` function.
+
+.. GENERATED FROM PYTHON SOURCE LINES 20-25
+
+Let's create a new feature class called `ChannelMean` that calculates the mean signal for each channel.
+We can optinally make it inherit from :class:`~features.NMFeature` but as long as it has an adequate constructor
+and a `calc_feature` method with the appropriate signatures it will work.
+The :func:`__init__` method should take the settings, channel names and sampling frequency as arguments.
+The `calc_feature` method should take the data and a dictionary of features as arguments and return the updated dictionary.
+
+.. GENERATED FROM PYTHON SOURCE LINES 25-58
+
+.. code-block:: Python
+
+    class ChannelMean:
+        def __init__(
+            self, settings: nm.NMSettings, ch_names: Iterable[str], sfreq: float
+        ) -> None:
+            # If required for feature calculation, store the settings,
+            # channel names and sampling frequency (optional)
+            self.settings = settings
+            self.ch_names = ch_names
+            self.sfreq = sfreq
+
+            # Here you can add any additional initialization code
+            # For example, you could store parameters for the functions\
+            # used in the calc_feature method
+
+            self.feature_name = "channel_mean"
+
+        def calc_feature(self, data: np.ndarray) -> dict:
+            # First, create an empty dictionary to store the calculated features
+            feature_results = {}
+
+            # Here you can add any feature calculation code
+            # This example simply calculates the mean signal for each channel
+            ch_means = np.mean(data, axis=1)
+
+            # Store the calculated features in the feature_results dictionary
+            # Be careful to use a unique keyfor each channel and metric you compute
+            for ch_idx, ch in enumerate(self.ch_names):
+                feature_results[f"{self.feature_name}_{ch}"] = ch_means[ch_idx]
+
+            # Return the updated feature_results dictionary to the stream
+            return feature_results
+
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 59-60
+
+Now we can instantiate settings and observe that the new feature has been added to the list of features
+
+.. GENERATED FROM PYTHON SOURCE LINES 60-66
+
[…]
