+++ _sources/logging.rst.txt
@@ -1,0 +1,5 @@
+logging
+=======
+
+.. autoclass:: utils.logging.NMLogger
+      :members:

[…]
