### 2025-06-02 neuromodulation/py_neuromodulation#392 @danielmk
@timonmerk I just pip installed again on a different machine but msgpack was still missing. I noticed that the default release installed with pip install py_neuromodulation is 0.0.5.

I was able to specifically install the new release with `pip install py_neuromodulation==0.1.3` but it requires Python>=3.11, which is contrary to the README that claims Python>=3.10

But the new 0.1.3 release does indeed install the missing msgpack that caused the initial error.
