### 2025-06-02 0d59861
Update README.rst Fix Python version
