+++ _sources/auto_examples/plot_1_example_BIDS.rst.txt
@@ -1,0 +1,914 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "auto_examples/plot_1_example_BIDS.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        :ref:`Go to the end <sphx_glr_download_auto_examples_plot_1_example_BIDS.py>`
+        to download the full example code.
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_auto_examples_plot_1_example_BIDS.py:
+
+
+ECoG Movement decoding example
+==============================
+
+.. GENERATED FROM PYTHON SOURCE LINES 8-17
+
+This example notebook read openly accessible data from the publication
+*Electrocorticography is superior to subthalamic local field potentials
+for movement decoding in Parkinson’s disease*
+(`Merk et al. 2022 <https://elifesciences.org/articles/75126>_`).
+The dataset is available `here <https://doi.org/10.7910/DVN/io.FLM>`_.
+
+For simplicity one example subject is automatically shipped within
+this repo at the *py_neuromodulation/data* folder, stored in
+`iEEG BIDS <https://www.nature.com/articles/s41597-019-0105-7>`_ format.
+
+.. GENERATED FROM PYTHON SOURCE LINES 19-24
+
+.. code-block:: Python
+
+    from sklearn import metrics, model_selection, linear_model
+    import matplotlib.pyplot as plt
+
+    import py_neuromodulation as nm
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 25-28
+
+Let's read the example using `mne_bids <https://mne.tools/mne-bids/stable/index.html>`_.
+The resulting raw object is of type `mne.RawArray <https://mne.tools/stable/generated/mne.io.RawArray.html>`_.
+We can use the properties such as sampling frequency, channel names, channel types all from the mne array and create the *nm_channels* DataFrame:
+
+.. GENERATED FROM PYTHON SOURCE LINES 28-57
+
+.. code-block:: Python
+
+
+    (
+        RUN_NAME,
+        PATH_RUN,
+        PATH_BIDS,
+        PATH_OUT,
+        datatype,
+    ) = nm.io.get_paths_example_data()
+
+    (
+        raw,
+        data,
+        sfreq,
+        line_noise,
+        coord_list,
+        coord_names,
+    ) = nm.io.read_BIDS_data(PATH_RUN=PATH_RUN)
+
+    channels = nm.utils.set_channels(
+        ch_names=raw.ch_names,
+        ch_types=raw.get_channel_types(),
+        reference="default",
+        bads=raw.info["bads"],
+        new_names="default",
+        used_types=("ecog", "dbs", "seeg"),
+        target_keywords=["MOV_RIGHT"],
+    )
+
+
+
+
+
+
+.. rst-class:: sphx-glr-script-out
+
+ .. code-block:: none
+
+    Extracting parameters from /opt/hostedtoolcache/Python/3.11.12/x64/lib/python3.11/site-packages/py_neuromodulation/data/sub-testsub/ses-EphysMedOff/ieeg/sub-testsub_ses-EphysMedOff_task-gripforce_run-0_ieeg.vhdr...
+    Setting channel info structure...
[…]

[…]
