### 2025-06-02 neuromodulation/py_neuromodulation#392 @timonmerk
Thanks again for the comment! I updated now the python version to 3.11 in the ReadMe.
Unfortunately I could not reproduce that the default pypi install is version 0.0.5. Would it be possible to share the installation command? I definitely want to set the default to the current release version.

I saw either with `uv add py-neuromodulation` or `uv pip install py-neuromodulation` version 0.1.3

> ➜  ins uv init
> Initialized project `ins`
> ➜  ins git:(master) ✗ uv venv
> Using CPython 3.12.6
> Creating virtual environment at: .venv
> Activate with: source .venv/bin/activate
> ➜  ins git:(master) ✗ source .venv/bin/activate
> (ins) ➜  ins git:(master) ✗ uv pip install py-neuromodulation
> Resolved 100 packages in 722ms
> Prepared 6 packages in 1.18s
> Installed 100 packages in 663ms
>  + annotated-types==0.7.0
>  + anyio==4.9.0
>  + bottle==0.13.3
>  + bounded-pool-executor==0.0.3
>  + category-encoders==2.8.1
>  + cbor2==5.6.5
>  + certifi==2025.4.26
>  + cftime==1.6.4.post1
>  + charset-normalizer==3.4.2
>  + click==8.2.1
>  + contourpy==1.3.2
>  + cycler==0.12.1
>  + decorator==5.2.1
>  + fastapi==0.115.12
>  + filelock==3.18.0
>  + fonttools==4.58.1
>  + fooof==1.1.1
>  + fsspec==2025.5.1
>  + future==1.0.0
>  + h11==0.16.0
>  + hf-xet==1.1.2
>  + httptools==0.6.4
>  + huggingface-hub==0.32.3
>  + idna==3.10
>  + imbalanced-learn==0.13.0
>  + jinja2==3.1.6
>  + joblib==1.5.1
>  + kiwisolver==1.4.8
>  + lazy-loader==0.4
>  + llvmlite==0.44.0
>  + markupsafe==3.0.2
>  + matplotlib==3.10.3
>  + mne==1.9.0
>  + mne-bids==0.16.0
>  + mne-connectivity==0.7.0
>  + mne-lsl==1.9.0
>  + mrmr-selection==0.2.8
>  + msgpack==1.1.0
>  + netcdf4==1.7.2
>  + nibabel==5.3.2
>  + nolds==0.6.1
>  + numba==0.61.2
>  + numpy==2.2.6
>  + packaging==25.0
>  + pandas==2.2.3
>  + patsy==1.0.1
>  + pillow==11.2.1
>  + platformdirs==4.3.8
>  + polars==1.30.0
>  + pooch==1.8.2
>  + pqdm==0.2.0
>  + proxy-tools==0.1.0
>  + psutil==7.0.0
>  + py-neuromodulation==0.1.3
>  + pyaml==25.5.0
>  + pyarrow==20.0.0
>  + pybispectra==1.2.1
>  + pydantic==2.11.5
>  + pydantic-core==2.33.2
>  + pyobjc-core==11.0
>  + pyobjc-framework-cocoa==11.0
>  + pyobjc-framework-quartz==11.0
>  + pyobjc-framework-security==11.0
>  + pyobjc-framework-webkit==11.0
>  + pyparrm==1.1.1
>  + pyparsing==3.2.3
>  + pyqt6==6.9.0
>  + pyqt6-qt6==6.9.0
>  + pyqt6-sip==13.10.2
>  + pyqtgraph==0.13.7
>  + python-dateutil==2.9.0.post0
>  + python-dotenv==1.1.0
>  + pytz==2025.2
>  + pywebview==5.4
>  + pyyaml==6.0.2
>  + qtpy==2.4.3
>  + requests==2.32.3
>  + scikit-learn==1.6.1
>  + scikit-optimize==0.10.2
>  + scipy==1.15.3
>  + seaborn==0.13.2
>  + setuptools==80.9.0
>  + six==1.17.0
>  + sklearn-compat==0.1.3
>  + skops==0.11.0
>  + sniffio==1.3.1
>  + starlette==0.46.2
>  + statsmodels==0.14.4
>  + tabulate==0.9.0
>  + threadpoolctl==3.6.0
>  + tqdm==4.67.1
>  + typing-extensions==4.14.0
>  + typing-inspection==0.4.1
>  + tzdata==2025.2
>  + urllib3==2.4.0
>  + uvicorn==0.34.3
>  + uvloop==0.21.0
>  + watchfiles==1.0.5
>  + websockets==15.0.1
>  + xarray==2025.4.0
> (ins) ➜  ins git:(master) ✗ uv pip show py-neuromodulation
> Name: py-neuromodulation
> Version: 0.1.3
> Location: /Users/Timon/Documents/else_desktop/ins/.venv/lib/python3.12/site-packages
> Requires: cbor2, fastapi, fooof, imbalanced-learn, joblib, llvmlite, matplotlib, mne, mne-bids, mne-connectivity, mne-lsl, mrmr-selection, msgpack, nibabel, nolds, numba, numpy, pandas, pyarrow, pybispectra, pydantic, pyparrm, pyqt6, pywebview, scikit-learn, scikit-optimize, scipy, seaborn, skops, uvicorn, uvloop
>
