### 2025-06-03 neuromodulation/py_neuromodulation#392 @danielmk
It works now, I'm getting 0.1.3

Thanks for looking into this. I'll close this issue now that pytest is running.
