### 2025-06-09 neuromodulation/py_neuromodulation#395 @danielmk
[JOSS Review] Checklist: GUI not working
This issue is part of the JOSS review process: https://github.com/openjournals/joss-reviews/issues/8258

It relates to the checklist point "Functionality".

`run_gui` works and I can open the GUI in google chrome as described in the documentation. However, when I click `OPEN FILE`, no file dialogue opens. Instead this error is raised.:
```
Traceback (most recent call last):
  File "C:\Users\Daniel\miniforge3\envs\pynmclean\Lib\site-packages\uvicorn\protocols\http\httptools_impl.py", line 409, in run_asgi
    result = await app(  # type: ignore[func-returns-value]
             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\Daniel\miniforge3\envs\pynmclean\Lib\site-packages\uvicorn\middleware\proxy_headers.py", line 60, in __call__
    return await self.app(scope, receive, send)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\Daniel\miniforge3\envs\pynmclean\Lib\site-packages\fastapi\applications.py", line 1054, in __call__
    await super().__call__(scope, receive, send)
  File "C:\Users\Daniel\miniforge3\envs\pynmclean\Lib\site-packages\starlette\applications.py", line 112, in __call__
    await self.middleware_stack(scope, receive, send)
  File "C:\Users\Daniel\miniforge3\envs\pynmclean\Lib\site-packages\starlette\middleware\errors.py", line 187, in __call__
    raise exc
  File "C:\Users\Daniel\miniforge3\envs\pynmclean\Lib\site-packages\starlette\middleware\errors.py", line 165, in __call__
    await self.app(scope, receive, _send)
  File "C:\Users\Daniel\miniforge3\envs\pynmclean\Lib\site-packages\starlette\middleware\exceptions.py", line 62, in __call__
    await wrap_app_handling_exceptions(self.app, conn)(scope, receive, send)
  File "C:\Users\Daniel\miniforge3\envs\pynmclean\Lib\site-packages\starlette\_exception_handler.py", line 53, in wrapped_app
    raise exc
  File "C:\Users\Daniel\miniforge3\envs\pynmclean\Lib\site-packages\starlette\_exception_handler.py", line 42, in wrapped_app
    await app(scope, receive, sender)
  File "C:\Users\Daniel\miniforge3\envs\pynmclean\Lib\site-packages\starlette\routing.py", line 714, in __call__
    await self.middleware_stack(scope, receive, send)
  File "C:\Users\Daniel\miniforge3\envs\pynmclean\Lib\site-packages\starlette\routing.py", line 734, in app
    await route.handle(scope, receive, send)
  File "C:\Users\Daniel\miniforge3\envs\pynmclean\Lib\site-packages\starlette\routing.py", line 288, in handle
    await self.app(scope, receive, send)
  File "C:\Users\Daniel\miniforge3\envs\pynmclean\Lib\site-packages\starlette\routing.py", line 76, in app
    await wrap_app_handling_exceptions(app, request)(scope, receive, send)
  File "C:\Users\Daniel\miniforge3\envs\pynmclean\Lib\site-packages\starlette\_exception_handler.py", line 53, in wrapped_app
    raise exc
  File "C:\Users\Daniel\miniforge3\envs\pynmclean\Lib\site-packages\starlette\_exception_handler.py", line 42, in wrapped_app
    await app(scope, receive, sender)
  File "C:\Users\Daniel\miniforge3\envs\pynmclean\Lib\site-packages\starlette\routing.py", line 73, in app
    response = await f(request)
               ^^^^^^^^^^^^^^^^
  File "C:\Users\Daniel\miniforge3\envs\pynmclean\Lib\site-packages\fastapi\routing.py", line 301, in app
    raw_response = await run_endpoint_function(
                   ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\Daniel\miniforge3\envs\pynmclean\Lib\site-packages\fastapi\routing.py", line 212, in run_endpoint_function
    return await dependant.call(**values)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\Daniel\miniforge3\envs\pynmclean\Lib\site-packages\py_neuromodulation\gui\backend\app_backend.py", line 259, in setup_offline_stream
    file_path=data["file_path"],
              ~~~~^^^^^^^^^^^^^
KeyError: 'file_path'
```

I am using google chrome on Windows 10. The pynm version is 0.1.3.
