### 2025-06-09 neuromodulation/py_neuromodulation#396 @danielmk
[JOSS Review] Checklist: Software paper
This issue is part of the JOSS review process: https://github.com/openjournals/joss-reviews/issues/8258

It relates to various checklist points regarding the [software paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.08258/joss.08258/10.21105.joss.08258.pdf).

I am not sure whether the paper accurately reflects the "State of the field". For example, references to some software packages that overlap in purpose are not in the paper. For example [here](https://github.com/mne-tools/mne-bids-pipeline/graphs/contributors) and [here](https://www.frontiersin.org/journals/neuroscience/articles/10.3389/fnins.2025.1522675/full). To properly reflect the state of the field you should briefly explain the commonalities and differences between your software and available software. If you feel like these do not warrant reference in the paper, you can explain here why. For example, if you feel like they don't sufficiently overlap in purpose. These two examples are not exhaustive, there might be other packages that warrant citation in the paper.

Also, I just noticed the cebra paper is not cited. Cebra appears in Figure 1 but not in the text. Also scikit-learn is not cited.

Two minor points:
The legend of Figure 1 should actually explain the figure. For example, what is even shown in b? The legend should at least explain what a and b are about.
Plotlly is a typo.
