+++ _sources/auto_examples/plot_3_example_sharpwave_analysis.rst.txt
@@ -1,0 +1,638 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "auto_examples/plot_3_example_sharpwave_analysis.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        :ref:`Go to the end <sphx_glr_download_auto_examples_plot_3_example_sharpwave_analysis.py>`
+        to download the full example code.
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_auto_examples_plot_3_example_sharpwave_analysis.py:
+
+
+Analyzing temporal features
+===========================
+
+.. GENERATED FROM PYTHON SOURCE LINES 8-32
+
+Time series data can be characterized using oscillatory components, but assumptions of sinusoidality are for real data rarely fulfilled.
+See *"Brain Oscillations and the Importance of Waveform Shape"* `Cole et al 2017 <https://doi.org/10.1016/j.tics.2016.12.008>`_ for a great motivation.
+We implemented here temporal characteristics based on individual trough and peak relations,
+based on the :meth:~`scipy.signal.find_peaks` method. The function parameter *distance* can be specified in the *settings.yaml*.
+Temporal features can be calculated twice for troughs and peaks. In the settings, this can be specified by setting *estimate* to true
+in *detect_troughs* and/or *detect_peaks*. A statistical measure (e.g. mean, max, median, var) can be defined as a resulting feature from the peak and
+trough estimates using the *apply_estimator_between_peaks_and_troughs* setting.
+
+In py_neuromodulation the following characteristics are implemented:
+
+.. note::
+    The nomenclature is written here for sharpwave troughs, but detection of peak characteristics can be computed in the same way.
+
+-  prominence:
+   :math:`V_{prominence} = |\frac{V_{peak-left} + V_{peak-right}}{2}| - V_{trough}`
+-  sharpness:
+   :math:`V_{sharpnesss} = \frac{(V_{trough} - V_{trough-5 ms}) + (V_{trough} - V_{trough+5 ms})}{2}`
+-  rise and decay rise time
+-  rise and decay steepness
+-  width (between left and right peaks)
+-  interval (between troughs)
+
+Additionally, different filter ranges can be parametrized using the *filter_ranges_hz* setting.
+Filtering is necessary to remove high frequent signal fluctuations, but limits also the true estimation of sharpness and prominence due to signal smoothing.
+
+.. GENERATED FROM PYTHON SOURCE LINES 32-45
+
+.. code-block:: Python
+
+
+    from typing import cast
+    import seaborn as sb
+    from matplotlib import pyplot as plt
+    from scipy import signal
+    from scipy.signal import fftconvolve
+    import numpy as np
+
+    import py_neuromodulation as nm
+    from py_neuromodulation import NMSettings
+    from py_neuromodulation.features import SharpwaveAnalyzer
+
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 46-47
+
+We will first read the example ECoG data and plot the identified features on the filtered time series.
+
+.. GENERATED FROM PYTHON SOURCE LINES 47-59
+
+.. code-block:: Python
+
+
+    RUN_NAME, PATH_RUN, PATH_BIDS, PATH_OUT, datatype = nm.io.get_paths_example_data()
+
+    (
+        raw,
+        data,
+        sfreq,
+        line_noise,
[…]
