### 2025-06-13 28e6a81
Adapt paper.md to address scientific need
