+++ paper/paper.bib
@@ -282,3 +282,32 @@ Publisher: Nature Publishing Group},
 	annote = {If you use this software, please cite it as below.},
 	file = {Zenodo Snapshot:C\:\\Users\\richa\\Zotero\\storage\\S9Z5AAHJ\\7979740.html:text/html},
 }
+
+@article{van_es_osl-ephys_2025,
+	title = {osl-ephys: a {Python} toolbox for the analysis of electrophysiology data},
+	volume = {Volume 19 - 2025},
+	issn = {1662-453X},
+	url = {https://www.frontiersin.org/journals/neuroscience/articles/10.3389/fnins.2025.1522675},
+	doi = {10.3389/fnins.2025.1522675},
+	abstract = {We describe OHBA Software Library for the analysis of electrophysiology data (osl-ephys). This toolbox builds on top of the widely used MNE-Python package and provides unique analysis tools for magneto−/electro-encephalography (M/EEG) sensor and source space analysis, which can be used modularly. In particular, it facilitates processing large amounts of data using batch parallel processing, with high standards for reproducibility through a config API and log keeping, and efficient quality assurance by producing HTML processing reports. It also provides new functionality for doing coregistration, source reconstruction and parcellation in volumetric space, allowing for an alternative pipeline that avoids the need for surface-based processing, e.g., through the use of Fieldtrip. Here, we introduce osl-ephys by presenting examples applied to a publicly available M/EEG data (the multimodal faces dataset). osl-ephys is open-source software distributed on the Apache License and available as a Python package through PyPi and GitHub.},
+	journal = {Frontiers in Neuroscience},
+	author = {van Es, Mats W. J. and Gohil, Chetan and Quinn, Andrew J. and Woolrich, Mark W.},
+	year = {2025},
+}
+
+@article {HBM:HBM23730,
+author = {Schirrmeister, Robin Tibor and Springenberg, Jost Tobias and Fiederer,
+  Lukas Dominique Josef and Glasstetter, Martin and Eggensperger, Katharina and Tangermann, Michael and
+  Hutter, Frank and Burgard, Wolfram and Ball, Tonio},
+title = {Deep learning with convolutional neural networks for EEG decoding and visualization},
+journal = {Human Brain Mapping},
+issn = {1097-0193},
+url = {http://dx.doi.org/10.1002/hbm.23730},
+doi = {10.1002/hbm.23730},
+month = {aug},
+year = {2017},
+keywords = {electroencephalography, EEG analysis, machine learning, end-to-end learning, brain–machine interface,
+  brain–computer interface, model interpretability, brain mapping},
+}
+
+
