+++ paper/paper.md
@@ -68,7 +68,7 @@ Analysis of electrophysiological recordings is commonly started by multiple prep
 
 # Statement of need
 
-The variety of different processing parameters poses a challenge to select optimal methods for machine learning-based decoding, and second, hinders reproducibility, since often necessary parameters are not reported in scientific literature or reproducible from the derived data. *py_neuromodulation* utilizes two parametrization files: *settings.yaml* and *channels.tsv*, that include all required steps for reproducibility. It allows for a streamlined pipeline of pre-processing, feature estimation and decoding. The performance contribution of various pre-processing steps can be assessed in combination with different feature estimation modalities and machine learning model architectures.  
+The variety of different processing parameters poses a challenge to select optimal methods for machine learning-based decoding, and second, hinders reproducibility, since often necessary parameters are not reported in scientific literature or reproducible from the derived data. While several toolboxes for processing and decoding using non-invasive neural recordings exists [@van_es_osl-ephys_2025;@HBM:HBM23730], there is a lack of analysis tools for invasive recordings specifically using local field potential recordings from either deep brain stimulation electrodes or electrocorticographical strip-electrodes. We aim to simply the pre-processing and feature-estimation using those invasive modalities. To simply this analysis, *py_neuromodulation* utilizes two parametrization files: *settings.yaml* and *channels.tsv*, that include all required steps for reproducibility. It allows for a streamlined pipeline of pre-processing, feature estimation and decoding. The performance contribution of various pre-processing steps can be assessed in combination with different feature estimation modalities and machine learning model architectures.  
 Offline analysis often includes whole-recording normalization, non-causal filtering, or feature computation with test-set data leakage [@merk_machine_2022], which limits real-time compatibility. In py_neuromodulation, the data-source is represented as a Python generator, and can be interchangeable for online and offline processing, meaning that there is no difference in handling data streamed directly from a brain implant, or stored from an archival dataset. This enables training of generalizable models that are directly applicable to online applications, using exactly the same algorithms. 
 Finally, electrophysiological research entails multiple technical and conceptual hurdles that pose substantial entry barriers for scientists without software development training. By providing an intuitive GUI for parametrization and raw-data and feature visualization, we aim to lower the domain knowledge requirements for entering this field.
 Py_neuromodulation was already used in several scientific projects with various applications: ECoG and DBS-LFP movement decoding [@merk_electrocorticography_2022;@cavallo_reinforcement_2025], movement intention classification [@kohler_dopamine_2024], seizure classification [@merk_invasive_2023], emotional valence decoding [@merk_invasive_2023], gait symptom estimation in rodents [@elgarulli_elgarullineurokin_2025], amongst others.
@@ -78,7 +78,7 @@ Py_neuromodulation was already used in several scientific projects with various 
 A data-source can be specified to be either a two-dimensional array (channels x time) read through [mne.io.read_raw](https://mne.tools/1.8/generated/mne.io.read_raw.html) supported recording standards or a [LabStreamingLayer](https://labstreaminglayer.org/#/) stream based on [mne-lsl](https://github.com/mne-tools/mne-lsl) (\autoref{fig:FigurePyNM}). If the data is passed as a simple numpy array, the sampling rate needs to be additionally specified. The parametrization **settings** and **channels** can be passed as additional optional parameters or automatically inferred.
[…]
