### 2025-06-15 neuromodulation/py_neuromodulation#396 @finsberg
I agree with @danielmk that you should elaborate more in the paper about how your software compares with other available softwares (ref "State of the field" in https://joss.readthedocs.io/en/latest/review_checklist.html#software-paper). This is important information for the reader to understand how your software compares to the other. I would like to see a new section in the paper about this. As a reference to other softwares you can also check out [this repo](https://github.com/openlists/ElectrophysiologySoftware)
