### 2025-06-19 #8258 @jromanowska
I can see that the review process is going well, great job, @finsberg , @danielmk , @timonmerk!
Just wanted to let you know that I'm on summer holidays entire July and won't be able to do much then, so if you need anything from me, next week is best time.
