### 2025-06-19 432faf5
Set Stream's verbose to be False by default + Only log to file if verbose
