+++ paper/paper.bib
@@ -310,4 +310,25 @@ keywords = {electroencephalography, EEG analysis, machine learning, end-to-end l
   brain–computer interface, model interpretability, brain mapping},
 }
 
+@misc{noauthor_openlistselectrophysiologysoftware_2025,
+	title = {openlists/{ElectrophysiologySoftware}},
+	url = {https://github.com/openlists/ElectrophysiologySoftware},
+	abstract = {A list of openly available software tools for (mostly human) electrophysiology.},
+	urldate = {2025-06-20},
+	publisher = {OpenLists},
+	month = jun,
+	year = {2025},
+	note = {original-date: 2018-07-19T00:01:45Z},
+	keywords = {awesome-list, data-analysis, ecog, eeg, electrophysiology, lfp, list, meg, open-science, open-source, software},
+}
 
+@article{GramfortEtAl2013a,
+  title = {{{MEG}} and {{EEG}} Data Analysis with {{MNE}}-{{Python}}},
+  author = {Gramfort, Alexandre and Luessi, Martin and Larson, Eric and Engemann, Denis A. and Strohmeier, Daniel and Brodbeck, Christian and Goj, Roman and Jas, Mainak and Brooks, Teon and Parkkonen, Lauri and H{\"a}m{\"a}l{\"a}inen, Matti S.},
+  year = {2013},
+  volume = {7},
+  pages = {1--13},
+  doi = {10.3389/fnins.2013.00267},
+  journal = {Frontiers in Neuroscience},
+  number = {267}
+}
