+++ paper/paper.md
@@ -66,9 +66,13 @@ By introducing py_neuromodulation, we aim to simplify and standardize the analys
 Recently, deep learning foundation models were presented that showed a performance leap on many clinical relevant downstream tasks [@yuan_brainwave_2024;@li_deep_2025]. Neural signal processing for machine learning and medical device engineering is a critical step for the advancement of the fields of neural decoding and brain computer interfaces. We, therefore, developed a neural signal processing, feature estimation and decoding toolbox [py_neuromodulation](https://github.com/neuromodulation/py_neuromodulation), that includes several established and standardized pipelines that simplify processing of neural recordings.  
 Analysis of electrophysiological recordings is commonly started by multiple preprocessing steps: resampling, notch-filtering, normalization, re-referencing (subtraction of neighboring channels), filtering (lowpass, highpass, bandpass), or artifact rejection [@cohen_analyzing_2014]. Next, the signals or features of interest are computed. Most commonly, signals are analyzed in the spectral domain. Here, different methods can be utilized, including Fast Fourier Transform, Welch transform, bandpass filtering, and others. In addition to spectral features, different feature modalities such as waveform-shape [@cole_brain_2017], periodic and aperiodic spectral parametrization [@donoghue_parameterizing_2020], burst features [@tinkhauser_beta_2018], and others can be computed.
 
+# State of the field
+
+Several toolboxes for pre-processing, feature estimation and decoding for neural recordings exist in the Python programming language [@noauthor_openlistselectrophysiologysoftware_2025]. Notably [mne-python](https://github.com/mne-tools/mne-python) is a general purpose toolbox for processing and analyzing MEG and EEG data [GramfortEtAl2013a]. [Braindecode](https://github.com/robintibor/braindecode) is an deep learning toolbox for end-to-end neural decoding and benchmarking [@HBM:HBM23730]. [OSL-Ephys](https://github.com/OHBA-analysis/osl-ephys) is a MEG and EEG toolbox for automatized preprocessing and source reconstruction [@van_es_osl-ephys_2025]. Several other tools implement computation of specific individual neural characteristics, such as [spectral parametrization](https://fooof-tools.github.io/fooof/index.html), [connectivity](https://mne.tools/mne-connectivity/stable/index.html) or [nonlinear dynamics](https://cschoel.github.io/nolds/). The previous mentioned neural signal processing and decoding tools are primarily optimized for analysis of non-invasive data. While many of the pre-processing and signal characterization steps remain similar in non-invasive recordings, there is a specific need to derive optimal analysis parameters for invasive recordings, for example from deep brain stimulation or electrocorticographical electrodes. Invasive local field potential recordings contain commonly signals of low number of channels, which results in signal processing without source reconstruction. Second, specific features are shown to be particularly relevant for subcortical recordings, such as bursts in different frequency bands  [@merk_electrocorticography_2022]. Therefore, there is a lack of analysis tools for invasive recordings that enable explorative neural data analysis and machine learning neural decoding based on local field potential recordings from either deep brain stimulation or stereotactic EEG electrodes.
+
 # Statement of need
 
[…]
