### 2025-06-20 88d60cd
add built gui
