+++ _downloads/eaa4305c75b19a1e2eea941f742a6331/plot_4_example_gridPointProjection.py
@@ -1,0 +1,207 @@
+"""
+Grid Point Projection
+=====================
+
+"""
+
+# %%
+# In ECoG datasets the electrode locations are usually different. For this reason, we established a grid
+# with a set of points defined in a standardized MNI brain.
+# Data is then interpolated to this grid, such that they are common across patients, which allows across patient decoding use cases.
+#
+# In this notebook, we will plot these grid points and see how the features extracted from our data can be projected into this grid space.
+#
+# In order to do so, we'll read saved features that were computed in the ECoG movement notebook.
+# Please note that in order to do so, when running the feature estimation, the settings
+#
+# .. note::
+#
+#     .. code-block:: python
+#
+#         stream.settings['postprocessing']['project_cortex'] = True
+#         stream.settings['postprocessing']['project_subcortex'] = True
+#
+#     need to be set to `True` for a cortical and/or subcortical projection.
+#
+
+# %%
+import numpy as np
+import matplotlib.pyplot as plt
+
+import py_neuromodulation as nm
+
+
+# %%
+# Read features from BIDS data
+# ----------------------------
+#
+# We first estimate features, with the `grid_point` projection settings enabled for cortex.
+
+
+# %%
+RUN_NAME, PATH_RUN, PATH_BIDS, PATH_OUT, datatype = nm.io.get_paths_example_data()
+
+(
+    raw,
+    data,
+    sfreq,
+    line_noise,
+    coord_list,
+    coord_names,
+) = nm.io.read_BIDS_data(PATH_RUN=PATH_RUN)
+
+settings = nm.NMSettings.get_fast_compute()
+
+settings.postprocessing.project_cortex = True
+
+channels = nm.utils.set_channels(
+    ch_names=raw.ch_names,
+    ch_types=raw.get_channel_types(),
+    reference="default",
+    bads=raw.info["bads"],
+    new_names="default",
+    used_types=("ecog", "dbs", "seeg"),
+    target_keywords=["MOV_RIGHT", "MOV_LEFT"],
+)
+
+stream = nm.Stream(
+    sfreq=sfreq,
+    channels=channels,
+    settings=settings,
+    line_noise=line_noise,
+    coord_list=coord_list,
+    coord_names=coord_names,
+    verbose=True,
+)
+
+features = stream.run(
+    data=data[:, : int(sfreq * 5)],
+    out_dir=PATH_OUT,
+    experiment_name=RUN_NAME,
+    save_csv=True,
+)
+
+# %%
+# From analysis.py, we use the :class:~`analysis.FeatureReader` class to load the data.
+
+# init analyzer
+feature_reader = nm.FeatureReader(feature_dir=PATH_OUT, feature_file=RUN_NAME)
+
+# %%
+# To perform the grid projection, for all computed features we check for every grid point if there is any electrode channel within the spatial range ```max_dist_mm```, and weight
+# this electrode contact by the inverse distance and normalize across all electrode distances within the maximum distance range.
+# This gives us a projection matrix that we can apply to streamed data, to transform the feature-channel matrix *(n_features, n_channels)* into the grid point matrix *(n_features, n_gridpoints)*.
+#
+# To save computation time, this projection matrix is precomputed before the real time run computation.
+# The cortical grid is stored in *py_neuromodulation/grid_cortex.tsv* and the electrodes coordinates are stored in *_space-mni_electrodes.tsv* in a BIDS dataset.
+#
+# .. note::
+#
+#     One remark is that our cortical and subcortical grids are defined for the **left** hemisphere of the brain and, therefore, electrode contacts are mapped to the left hemisphere.
+#
[…]
