+++ _sources/auto_examples/sg_execution_times.rst.txt
@@ -1,0 +1,61 @@
+
+:orphan:
+
+.. _sphx_glr_auto_examples_sg_execution_times:
+
+
+Computation times
+=================
+**00:50.931** total execution time for 9 files **from auto_examples**:
+
+.. container::
+
+  .. raw:: html
+
+    <style scoped>
+    <link href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/5.3.0/css/bootstrap.min.css" rel="stylesheet" />
+    <link href="https://cdn.datatables.net/1.13.6/css/dataTables.bootstrap5.min.css" rel="stylesheet" />
+    </style>
+    <script src="https://code.jquery.com/jquery-3.7.0.js"></script>
+    <script src="https://cdn.datatables.net/1.13.6/js/jquery.dataTables.min.js"></script>
+    <script src="https://cdn.datatables.net/1.13.6/js/dataTables.bootstrap5.min.js"></script>
+    <script type="text/javascript" class="init">
+    $(document).ready( function () {
+        $('table.sg-datatable').DataTable({order: [[1, 'desc']]});
+    } );
+    </script>
+
+  .. list-table::
+   :header-rows: 1
+   :class: table table-striped sg-datatable
+
+   * - Example
+     - Time
+     - Mem (MB)
+   * - :ref:`sphx_glr_auto_examples_plot_7_lsl_example.py` (``plot_7_lsl_example.py``)
+     - 00:23.350
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_plot_0_first_demo.py` (``plot_0_first_demo.py``)
+     - 00:12.493
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_plot_1_example_BIDS.py` (``plot_1_example_BIDS.py``)
+     - 00:08.190
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_plot_4_example_gridPointProjection.py` (``plot_4_example_gridPointProjection.py``)
+     - 00:02.666
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_plot_2_example_add_feature.py` (``plot_2_example_add_feature.py``)
+     - 00:02.219
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_plot_3_example_sharpwave_analysis.py` (``plot_3_example_sharpwave_analysis.py``)
+     - 00:01.706
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_plot_6_real_time_demo.py` (``plot_6_real_time_demo.py``)
+     - 00:00.305
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_plot_8_cebra_example.py` (``plot_8_cebra_example.py``)
+     - 00:00.003
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_plot_5_example_rmap_computing.py` (``plot_5_example_rmap_computing.py``)
+     - 00:00.000
+     - 0.0
