### 2025-06-20 neuromodulation/py_neuromodulation#395 @timonmerk
So sorry @danielmk. The issue was fixed, but we forgot to upload the newly built js frontend files. Since the app was started in `debug` False mode, the previous bundled files were used. I've uploaded the fixed files now and merged the `main` branch also in `paper` now. 88d60cda5ed9032ff124a1ddc654bdcae9623a6f

For an example file you can just navigate to py_neuromodulation/data/sub-testsub/ses-EphysMedOff/iEEG/ and select the example .vhdr file (sub-testsub_ses-EphysMedOff_task-gripforce_run-0_ieeg.vhdr).

Thanks again for bringing this up!
