### 2025-06-27 neuromodulation/py_neuromodulation#395 @danielmk
I am now able to open the file explorer in the GUI but it does not actually show the files in the directory and instead raises an HTTP error, see screenshot. Same problem in chrome and edge.

![Image](https://github.com/user-attachments/assets/c8f9a014-5bc2-4e5e-95bc-1f2a27f2c12d)
