### 2025-07-03 8ce2eb4
Simple more pythonic code
