### 2025-07-10 fb07559
add test for different filter ranges
