### 2025-07-10 ce4c417
add swarpwave test for multiple channels
