### 2025-07-10 c7f7cfa
prevent parallel compute for tests
