### 2025-07-10 6317a4a
Update tests.yml - Remove macos testing
