+++ .github/workflows/tests.yml
@@ -28,8 +28,8 @@ jobs:
         platform:
           - os: ubuntu-latest
             name: Linux
-          - os: macos-latest
-            name: MacOS
+#          - os: macos-latest
+#            name: MacOS
           - os: windows-latest
             name: Windows
         python:
