### 2025-07-29 #8258 @timonmerk
Hey @jromanowska, thanks for getting back about our publication. We're actively working on an issue raised by @danielmk https://github.com/neuromodulation/py_neuromodulation/issues/395, which we aim to resolve this week.
Apart from that, we already adapted the paper https://github.com/neuromodulation/py_neuromodulation/commit/c0e01f7f426beca2f92cd95f6b054b6b283358cd and addressed different issues raised by the reviewers
