+++ .github/workflows/python-publish.yml
@@ -16,7 +16,7 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v4
         with:
-          python-version: '3.11'  # Specify your Python version
+          python-version: '3.12'  # Specify your Python version
 
       - name: Install dependencies
         run: |
