+++ paper/paper.md
@@ -25,6 +25,8 @@ authors:
   - name: Thomas S. Binns
     orcid: 0000-0003-0657-0891
     affiliation: "1, 2, 3"
+  - naem: Ahmed Tarek Kamel Abdalfatah
+    affilitation: "1"
   - name: Alessia Cavallo
     orcid: 0000-0003-1717-1378
     affiliation: "1, 2"
