+++ paper/paper.md
@@ -25,7 +25,7 @@ authors:
   - name: Thomas S. Binns
     orcid: 0000-0003-0657-0891
     affiliation: "1, 2, 3"
-  - naem: Ahmed Tarek Kamel Abdalfatah
+  - name: Ahmed Tarek Kamel Abdalfatah
     affilitation: "1"
   - name: Alessia Cavallo
     orcid: 0000-0003-1717-1378
