### 2025-07-31 e83cd5d
Update paper.md
