### 2025-08-01 #8258 @jromanowska
Thank you, @danielmk! I can see that the checklist is also completed for the other reviewer. @finsberg - please confirm that the submission is satisfactory now.
