### 2025-08-01 82e04f4
Update paper.md
Small corrections to ease the flow of text
