### 2025-08-01 834ebe8
add paper suggestions
