+++ paper/paper.md
@@ -61,7 +61,7 @@ bibliography: paper.bib
 # Summary
 
 Invasive brain signal decoding can revolutionize the clinical utility of neurotechnological therapies. Potential signal sources stem from electroencephalography (EEG), electrocorticography (ECoG) and local field potentials (LFP) recorded from deep brain stimulation (DBS) electrodes. The application of machine learning methods to these signals requires pre-processing and feature extraction – complex analyses, often lacking standardization and clear documentation. Here, we introduce [py_neuromodulation](https://github.com/neuromodulation/py_neuromodulation), a toolbox designed for standardized signal processing, feature extraction, and decoding of electrophysiological data. All parameters are explicitly defined in dedicated settings and channel parameterization files. The framework processes both data streamed in real-time and from recordings on disk using the same pipeline. Additionally, a browser-based graphical user interface (GUI) enables intuitive usage and visualization of the processing pipeline without requiring any code modification.  
-By introducing py_neuromodulation, we aim to simplify and standardize the analysis of electrophysiological recordings, facilitating reproducibility and accessibility in the field of brain signal decoding. Our tool bridges the fields of neuroscience and neural engineering, providing machine learning and neurotechnology researchers with reproducible methods for the development of generalizable machine learning algorithms. 
+By introducing `py_neuromodulation`, we aim to simplify and standardize the analysis of electrophysiological recordings, facilitating reproducibility and accessibility in the field of brain signal decoding. Our tool bridges the fields of neuroscience and neural engineering, providing machine learning and neurotechnology researchers with reproducible methods for the development of generalizable machine learning algorithms. 
 
 # Introduction
 
@@ -74,10 +74,10 @@ Several toolboxes for pre-processing, feature estimation and decoding for neural
 
 # Statement of need
 
-We aim to simply the pre-processing and feature-estimation using those invasive modalities. To simply this analysis, *py_neuromodulation* utilizes two parametrization files: *settings.yaml* and *channels.tsv*, that include all required steps for reproducibility. It allows for a streamlined pipeline of pre-processing, feature estimation and decoding. The performance contribution of various pre-processing steps can be assessed in combination with different feature estimation modalities and machine learning model architectures.  
-Offline analysis often includes whole-recording normalization, non-causal filtering, or feature computation with test-set data leakage [@merk_machine_2022], which limits real-time compatibility. In py_neuromodulation, the data-source is represented as a Python generator, and can be interchangeable for online and offline processing, meaning that there is no difference in handling data streamed directly from a brain implant, or stored from an archival dataset. This enables training of generalizable models that are directly applicable to online applications, using exactly the same algorithms. 
+We aim to simply the pre-processing and feature-estimation using those invasive modalities. To simply this analysis, `py_neuromodulation` utilizes two parametrization files: *settings.yaml* and *channels.tsv*, that include all required steps for reproducibility. It allows for a streamlined pipeline of pre-processing, feature estimation and decoding. This enables the assessment of performance contributions of various pre-processing steps in combination with feature estimation modalities and machine learning model architectures for different neural signal processing and decoding applications.
+Offline analysis often includes whole-recording normalization, non-causal filtering, or feature computation with test-set data leakage [@merk_machine_2022], which limits real-time compatibility. In `py_neuromodulation`, the data-source is represented as a Python generator, and can be interchangeable for online and offline processing, meaning that there is no difference in handling data streamed directly from a brain implant, or stored from an archival dataset. This enables training of generalizable models that are directly applicable to online applications, using exactly the same algorithms. 
[…]
