### 2025-08-04 #8258 @timonmerk
I've created a release with version v0.1.5 (https://github.com/neuromodulation/py_neuromodulation/releases/tag/v0.1.5) and here the zenodo doi: [10.5281/zenodo.15115194](https://doi.org/10.5281/zenodo.15115194).
