### 2025-08-05 #8258 @jromanowska
Please, correct the affiliations on Zenodo, so that they are exactly the same as in the paper (simply edit that in Zenodo).

Also: "Antonio Brotons" in the paper should probably be "Antonio Martinez Brotons", like in Zenodo?
