### 2025-08-05 #8258 @timonmerk
@jromanowska I've fixed the affiliations in zenodo: https://zenodo.org/records/16739068. Many thanks!
