### 2025-08-07 6822daf
Fixed Fooof zero bug; reference avarage bug; plus added unit tests
