+++ _sources/auto_examples/plot_7_lsl_example.rst.txt
@@ -1,0 +1,305 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "auto_examples/plot_7_lsl_example.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        :ref:`Go to the end <sphx_glr_download_auto_examples_plot_7_lsl_example.py>`
+        to download the full example code.
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_auto_examples_plot_7_lsl_example.py:
+
+
+Lab Streaming Layer (LSL) Example
+=================================
+
+This toolbox implements the lsl ecosystem which can be utilized for offline use cases as well as live streamings
+In this example the data introduced in the first demo is being analyzed
+in a similar manner, This time however integrating an lsl stream.
+
+.. GENERATED FROM PYTHON SOURCE LINES 12-15
+
+.. code-block:: Python
+
+    from matplotlib import pyplot as plt
+    import py_neuromodulation as nm
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 16-17
+
+Let’s get the example data from the provided BIDS dataset and create the channels DataFrame.
+
+.. GENERATED FROM PYTHON SOURCE LINES 17-130
+
+.. code-block:: Python
+
+
+    if __name__ == "__main__":
+        (
+            RUN_NAME,
+            PATH_RUN,
+            PATH_BIDS,
+            PATH_OUT,
+            datatype,
+        ) = nm.io.get_paths_example_data()
+
+        (
+            raw,
+            data,
+            sfreq,
+            line_noise,
+            coord_list,
+            coord_names,
+        ) = nm.io.read_BIDS_data(PATH_RUN=PATH_RUN)
+
+        channels = nm.utils.set_channels(
+            ch_names=raw.ch_names,
+            ch_types=raw.get_channel_types(),
+            reference="default",
+            bads=raw.info["bads"],
+            new_names="default",
+            used_types=("ecog", "dbs", "seeg"),
+            target_keywords=["MOV_RIGHT"],
+        )
+
+        # %%
+        # Playing the Data
+        # ----------------
+        #
+        # Now we need our data to be represeted in the LSL stream.
+        # For this example an mne_lsl.Player is utilized, which is playing our earlier
+        # recorded data. However, you could make use of any LSL source (live or
+        # offline).
+        # If you want to bind your own data source, make sure to specify the
+        # necessary parameters (data type, type, name) accordingly.
+        # If you are unsure about the parameters of your data source you can
+        # always search for available lsl streams.
+        #
+
+        settings = nm.NMSettings.get_fast_compute()
+
+        player = nm.stream.LSLOfflinePlayer(raw=raw, stream_name="example_stream")
+
+        player.start_player(chunk_size=30, n_repeat=1)
+        import time
+
+        time.sleep(5)
+        # %%
+        # Creating the LSLStream object
+        # -----------------------------
+        #
+        # Next let’s create a Stream analog to the First Demo’s example However as
+        # we run the stream, we will set the *lsl-stream* value to True and pass
+        # the stream name we earlier declared when initializing the player object.
+
+        settings.features.welch = False
+        settings.features.fft = True
+        settings.features.bursts = False
[…]
