### 2025-08-13 905633c
JOSS submission edits
