### 2025-08-13 #8258 @timonmerk
Thanks @Kevin-Mattheus-Moerman, I've addressed the issues. Here the Zenodo Repo https://zenodo.org/records/16739068 and the updated paper: https://github.com/neuromodulation/py_neuromodulation/actions/runs/16947896562
