### 2025-08-14 #8258 @timonmerk
@Kevin-Mattheus-Moerman and @jromanowska one final question. Currently the language tags are set to Matlab and JavaScript. https://joss.theoj.org/papers/10.21105/joss.08258 Would it be possible to change that to Python and JavaScript? Our package doesn't use MatLab
