### 2025-08-14 1595806
add JOSS paper reference in ReadME
