+++ docs/source/index.rst
@@ -51,6 +51,15 @@ The original intention for writing this toolbox was movement decoding from invas
 The application however could be any neural decoding and analysis problem.
 *py_neuromodulation* offers wrappers around common practice machine learning methods for efficient analysis.
 
+
+Journal of Open Source Software publication
+-------------------------------------------
+
+reStructuredText:
+.. image:: https://joss.theoj.org/papers/10.21105/joss.08258/status.svg
+   :target: https://doi.org/10.21105/joss.08258
+
+
 References
 ----------
 
