+++ README.rst
@@ -1,5 +1,14 @@
+
+
 py_neuromodulation
 ==================
+
+Journal of Open Source Science publication:
+
+.. image:: https://joss.theoj.org/papers/10.21105/joss.08258/status.svg
+   :target: https://doi.org/10.21105/joss.08258
+
+
 
 Documentation: https://neuromodulation.github.io/py_neuromodulation/
 
@@ -82,14 +91,6 @@ Contact information
 -------------------
 For any question or suggestion please find my contact
 information at `my GitHub profile <https://github.com/timonmerk>`_.
-
-Journal of Open Source Software publication
--------------------------------------------
-
-reStructuredText:
-.. image:: https://joss.theoj.org/papers/10.21105/joss.08258/status.svg
-   :target: https://doi.org/10.21105/joss.08258
-
 
 Contributing guide
 ------------------
