### 2025-08-14 dfce248
JOSS documentation
