+++ _downloads/3b4900a2b2818ff30362215b76f7d5eb/plot_1_example_BIDS.py
@@ -1,0 +1,225 @@
+"""
+ECoG Movement decoding example
+==============================
+
+"""
+
+# %%
+# This example notebook read openly accessible data from the publication
+# *Electrocorticography is superior to subthalamic local field potentials
+# for movement decoding in Parkinson’s disease*
+# (`Merk et al. 2022 <https://elifesciences.org/articles/75126>_`).
+# The dataset is available `here <https://doi.org/10.7910/DVN/io.FLM>`_.
+#
+# For simplicity one example subject is automatically shipped within
+# this repo at the *py_neuromodulation/data* folder, stored in
+# `iEEG BIDS <https://www.nature.com/articles/s41597-019-0105-7>`_ format.
+
+# %%
+from sklearn import metrics, model_selection, linear_model
+import matplotlib.pyplot as plt
+
+import py_neuromodulation as nm
+
+# %%
+# Let's read the example using `mne_bids <https://mne.tools/mne-bids/stable/index.html>`_.
+# The resulting raw object is of type `mne.RawArray <https://mne.tools/stable/generated/mne.io.RawArray.html>`_.
+# We can use the properties such as sampling frequency, channel names, channel types all from the mne array and create the *nm_channels* DataFrame:
+
+(
+    RUN_NAME,
+    PATH_RUN,
+    PATH_BIDS,
+    PATH_OUT,
+    datatype,
+) = nm.io.get_paths_example_data()
+
+(
+    raw,
+    data,
+    sfreq,
+    line_noise,
+    coord_list,
+    coord_names,
+) = nm.io.read_BIDS_data(PATH_RUN=PATH_RUN)
+
+channels = nm.utils.set_channels(
+    ch_names=raw.ch_names,
+    ch_types=raw.get_channel_types(),
+    reference="default",
+    bads=raw.info["bads"],
+    new_names="default",
+    used_types=("ecog", "dbs", "seeg"),
+    target_keywords=["MOV_RIGHT"],
+)
+
+
+# %%
+# This example contains the grip force movement traces, we'll use the *MOV_RIGHT* channel as a decoding target channel.
+# Let's check some of the raw feature and time series traces:
+
+plt.figure(figsize=(12, 4), dpi=300)
+plt.subplot(121)
+plt.plot(raw.times, data[-1, :])
+plt.xlabel("Time [s]")
+plt.ylabel("a.u.")
+plt.title("Movement label")
+plt.xlim(0, 20)
+
+plt.subplot(122)
+for idx, ch_name in enumerate(channels.query("used == 1").name):
+    plt.plot(raw.times, data[idx, :] + idx * 300, label=ch_name)
+plt.legend(bbox_to_anchor=(1, 0.5), loc="center left")
+plt.title("ECoG + STN-LFP time series")
+plt.xlabel("Time [s]")
+plt.ylabel("Voltage a.u.")
+plt.xlim(0, 20)
+
+# %%
+settings = nm.NMSettings.get_fast_compute()
+
+settings.features.welch = True
+settings.features.fft = True
+settings.features.bursts = True
+settings.features.sharpwave_analysis = True
+settings.features.coherence = True
+
+settings.coherence_settings.channels = [["LFP_RIGHT_0", "ECOG_RIGHT_0"]]
+
+settings.coherence_settings.frequency_bands = ["high beta", ]
+settings.sharpwave_analysis_settings.estimator["mean"] = []
+settings.sharpwave_analysis_settings.sharpwave_features.enable_all()
+for sw_feature in settings.sharpwave_analysis_settings.sharpwave_features.list_all():
+    settings.sharpwave_analysis_settings.estimator["mean"].append(sw_feature)
+
+# %%
+stream = nm.Stream(
[…]
