+++ _sources/api_documentation.rst.txt
@@ -1,0 +1,84 @@
+API Documentation
+=================
+
+
+Analysis
+--------
+
+.. toctree::
+   :maxdepth: 4
+
+   decode
+   feature_reader
+   plots
+   RMAP
+   stats
+
+
+Feature Classes
+---------------
+
+.. toctree::
+   :maxdepth: 4
+
+   bispectra
+   bursts
+   coherence
+   fooof
+   hjorth
+   linelength
+   mne_connectivity
+   nolds
+   oscillatory
+   sharpwaves
+   bandpower
+
+
+Filter
+------
+
+.. toctree::
+   :maxdepth: 4
+
+   mne_filter
+   notch_filter
+
+
+Processing
+----------
+
+.. toctree::
+   :maxdepth: 4
+
+   artifacts
+   normalization
+   projection
+   rereference
+   resample
+
+
+stream
+------
+
+.. toctree::
+   :maxdepth: 4
+
+   generator
+   mnelsl_player
+   mnelsl_stream
+   settings
+   stream
+
+
+utils
+-----
+
+.. toctree::
+   :maxdepth: 4
+
+   channels
+   database
+   io
+   logging
+   types
+
