### 2025-09-18 bdfe2cd
Creates a main.py that helps in creating the dmg
