+++ py_neuromodulation/default_settings.yaml
@@ -54,10 +54,10 @@ raw_normalization_settings:
   clip: 3
 
 preprocessing_filter:
-  bandstop_filter: true
-  lowpass_filter: true
-  highpass_filter: true
-  bandpass_filter: true
+  bandstop_filter: false
+  lowpass_filter: false
+  highpass_filter: false
+  bandpass_filter: false
   bandstop_filter_settings: [100, 160] # [low_hz, high_hz]
   bandpass_filter_settings: [3, 200] # [hz, _hz]
   lowpass_filter_cutoff_hz: 200
@@ -245,3 +245,4 @@ bispectrum_settings:
     mean: true
     sum: true
     var: true
+
