### 2025-11-04 1c4acd9
Fix Pydantic model_fields deprecation warning
