### 2025-11-04 cb6fab4
Update uv setup Github action
