### 2025-11-04 5585635
Remove redundant env activation
