### 2025-11-04 4fef588
Update GitHub action (#411)
Update uv setup Github action
